+++ lib/informers/utils/core.rb
@@ -3,5 +3,9 @@ module Informers
     def self.dispatch_callback(progress_callback, data)
       progress_callback.(data) if progress_callback
     end
+
+    def self.calculate_reflect_offset(i, w)
+      ((i + w) % (2 * w) - w).abs
+    end
   end
 end